Split invoices_record models
